move the run and other active code buttons to the bottom
Students commented that it is difficult to click run, then quickly
scroll down past the code to the output to see it. Sometimes this is
an issue because input() triggers an alerts, which block scrolling.
Other times, it is an issue because turtles begin moving and we don't
get to see the very beginning of the motion.